backend/pyproject.toml
[tool.poetry]
name = "you do too backend"
version = "0.1.0"
description = ""
authors = ["Александр <av.sanya@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.11.7"
sqlalchemy = "^2.0.41"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
asyncpg = "^0.30.0"
psycopg = "^3.2.9"
alembic = "^1.16.4"
pydantic-settings = "^2.10.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
